refactor(fold): wire hasFoldingWin to Rust implementation
Replace the C hasFoldingWin implementation with a thin wrapper that
calls the Rust rs_hasFoldingWin function. The Rust implementation
returns a FoldingResult struct containing all the needed output values.

Also fix LineNr type mismatch: Rust was using i64 but C linenr_T is
int32_t. This caused struct layout issues and memory corruption.

+++ src/nvim-rs/fold/src/lib.rs
@@ -14,7 +14,7 @@ use std::ffi::{c_char, c_int};
 use nvim_window::WinHandle;
 
 /// Line number type (matches linenr_T in C).
-pub type LineNr = i64;
+pub type LineNr = i32;
 
 /// Result struct for hasFoldingWin.
 ///
